MAKE `CrudModelActionTestCase` for extending unit tests
