+++ tests/Unit/CrudModelActionCreateTest.php
@@ -6,37 +6,39 @@ namespace Sfneal\CrudModelActions\Tests\Unit;
 
 use Exception;
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Http\Request;
 use Illuminate\Http\Response;
-use Illuminate\Support\Facades\Event;
 use Sfneal\CrudModelActions\Tests\Assets\Actions\CreatePeopleAction;
-use Sfneal\CrudModelActions\Tests\Assets\Events\MockTrackingEvent;
-use Sfneal\CrudModelActions\Tests\Assets\Requests\PostRequest;
-use Sfneal\CrudModelActions\Tests\TestCase;
+use Sfneal\CrudModelActions\Tests\CrudModelActionTestCase;
 use Sfneal\Testing\Models\People;
-use Sfneal\Testing\Utils\Interfaces\RequestCreator;
-use Sfneal\Testing\Utils\Traits\EventFaker;
 
-class CrudModelActionCreateTest extends TestCase implements RequestCreator
+class CrudModelActionCreateTest extends CrudModelActionTestCase
 {
-    use EventFaker;
-    use PostRequest;
+    /**
+     * Execute a CrudModelAction.
+     *
+     * @return Response
+     * @throws Exception
+     */
+    protected function executeAction(): Response
+    {
+        return (new CreatePeopleAction($this->request))->execute();
+    }
 
     /**
-     * Retrieve a post request with mock data.
+     * Retrieve an array of data to be added to the mock request.
      *
-     * @return Request
+     * @return array
      */
-    private function postRequest(): Request
+    protected function requestData(): array
     {
-        return $this->createRequest([], [
+        return [
             'data' => [
                 'name_first' => 'Victor',
                 'name_last' => 'Hedman',
                 'email' => 'vhedman@example.com',
                 'age' => 29,
             ]
-        ]);
+        ];
     }
 
     /**
@@ -45,14 +47,13 @@ class CrudModelActionCreateTest extends TestCase implements RequestCreator
      */
     public function model_can_be_created()
     {
-        $request = $this->postRequest();
-        $response = (new CreatePeopleAction($request))->execute();
+        $this->executeAction();
 
         $query = People::query()
-            ->where('name_first', '=', $request->input('data.name_first'))
-            ->where('name_last', '=', $request->input('data.name_last'))
-            ->where('email', '=', $request->input('data.email'))
-            ->where('age', '=', $request->input('data.age'));
+            ->where('name_first', '=', $this->request->input('data.name_first'))
+            ->where('name_last', '=', $this->request->input('data.name_last'))
+            ->where('email', '=', $this->request->input('data.email'))
+            ->where('age', '=', $this->request->input('data.age'));
 
         $this->assertSame(1, $query->count());
 
@@ -62,35 +63,8 @@ class CrudModelActionCreateTest extends TestCase implements RequestCreator
         $this->assertInstanceOf(Model::class, $model);
         $this->assertInstanceOf(People::class, $model);
 
-        foreach ($request->input('data') as $key => $attribute) {
+        foreach ($this->request->input('data') as $key => $attribute) {
             $this->assertEquals($attribute, $model->{$key});
         }
-    }
-
-    /**
-     * @test
-     * @throws Exception
-     */
-    public function action_fires_event()
-    {
-        $this->eventFaker();
-        $request = $this->postRequest();
-        $response = (new CreatePeopleAction($request))->execute();
-
-        Event::assertDispatched(MockTrackingEvent::class);
-    }
-
-    /**
-     * @test
-     * @throws Exception
-     */
-    public function action_return_response()
-    {
-        $request = $this->postRequest();
-        $response = (new CreatePeopleAction($request))->execute();
-
-        $this->assertTrue(session()->has('success'));
-        $this->assertIsString(session('success'));
-        $this->assertInstanceOf(Response::class, $response);
     }
 }
